add new statistics modules and tests
