+++ m3c2/pipeline/param_manager.py
@@ -5,7 +5,7 @@ from typing import Tuple
 import numpy as np
 
 from m3c2.config.pipeline_config import PipelineConfig
-from m3c2.statistics import StatisticsService
+from m3c2.statistics.distance_stats import _load_params
 
 logger = logging.getLogger(__name__)
 
@@ -40,7 +40,7 @@ class ParamManager:
             params_path = os.path.join(
                 out_base, f"{cfg.process_python_CC}_{tag}_m3c2_params.txt"
             )
-            normal, projection = StatisticsService._load_params(params_path)
+            normal, projection = _load_params(params_path)
             if not np.isnan(normal) and not np.isnan(projection):
                 logger.info(
                     "[Params] geladen: %s (NormalScale=%.6f, SearchScale=%.6f)",
