+++ tests/test_pipeline/test_component_factory.py
@@ -10,7 +10,6 @@ from m3c2.pipeline.component_factory import PipelineComponentFactory
 from m3c2.importer.data_loader import DataLoader
 from m3c2.m3c2_core.param_handler.scale_estimator import ScaleEstimator
 from m3c2.m3c2_core.m3c2_executor import M3C2Executor
-from archive.outlier_handler import OutlierHandler
 from m3c2.statistics.statistics_runner import StatisticsRunner
 from m3c2.visualization.services.visualization_runner import VisualizationRunner
 
@@ -32,7 +31,7 @@ def test_factory_creates_configured_components():
     assert scale_estimator.strategy_name == "radius"
 
     assert isinstance(factory.create_m3c2_executor(), M3C2Executor)
-    # assert isinstance(factory.create_outlier_handler(), OutlierHandler)
+    # assert isinstance(factory.create_outlier_handler(), object)
 
     stats_runner = factory.create_statistics_runner()
     assert isinstance(stats_runner, StatisticsRunner)
